+++ LMeter/ACT/DataStructures/Encounter.cs
@@ -15,9 +15,9 @@ public class Encounter : IActData<Encounter>
     private static readonly Dictionary<string, MemberInfo> _textTagMembers = 
         typeof(Encounter).GetMembers().Where(x => Attribute.IsDefined(x, typeof(TextTagAttribute))).ToDictionary((x) => x.Name.ToLower());
 
-    public string GetFormattedString(string format, string numberFormat)
+    public string GetFormattedString(string format, string numberFormat, bool emptyIfZero)
     {
-        return TextTagFormatter.TextTagRegex.Replace(format, new TextTagFormatter(this, numberFormat, _textTagMembers).Evaluate);
+        return TextTagFormatter.TextTagRegex.Replace(format, new TextTagFormatter(this, numberFormat, emptyIfZero, _textTagMembers).Evaluate);
     }
 
     [TextTag]